feat(twins): M1 closeout sweep — split the engine server module, zero mirror language outside shared-types (F-685)
M6 (published contract, merged to main mid-flight) already did most of
this ticket's work: the mirror-parity scripts are deleted, the
copy-marker gate is hard-blocking with an EMPTY allowlist (stronger than
the planned 2-entry one), and redaction is centralized in
@pome-sh/shared-types/redaction. What remained on top of the three
ports:

- packages/sdk/src/server.ts split: boot/dispatch helpers + ok/created
  move to server-helpers.ts (file-size gate; TwinBootError and
  isLoopbackHost stay re-exported from @pome-sh/sdk/server) — replaces
  the temporary file-size headers from the F-682/F-684 lanes
- mirror-marker vocabulary swept to zero outside packages/shared-types
  (whose five 'byte-identical' OTel usages are determinism domain
  vocabulary in a change-gated package — ruled on F-685, cross-repo
  reword offered as follow-up)
- stale doc rows: CONTRACT.md Boot arrangement now names the shipped
  sdk dist; twin-slack README's auth row points at the engine instead
  of the deleted src/auth.ts

Integration gauntlet on this branch: contract 64 (61 pass / 3 per-twin
skips), 1166 workspace tests, cli 541/541 via the cli-ci tarball flow,
typecheck clean, copy-marker/code-health/no-cloud-imports/knip green.
